feat: Add stored procedure for adding bonus correction
The code changes include the addition of a new stored procedure called `AddBonus` in the `6-bonus.sql` file. This procedure allows for the addition of a new correction for a student by specifying the user ID, project name, and score. If the project does not exist in the `projects` table, it is inserted first. Then, the correction is inserted into the `corrections` table with the corresponding user ID, project ID, and score.